+++ Engine/ac/dynobj/cc_agsdynamicobject.cpp
@@ -55,52 +55,52 @@ int AGSCCDynamicObject::UnserializeInt() {
     return *iptr;
 }
 
-void AGSCCDynamicObject::Read(const char *address, int offset, void *dest, int size)
+void AGSCCDynamicObject::Read(const char *address, intptr_t offset, void *dest, int size)
 {
     memcpy(dest, address + offset, size);
 }
 
-uint8_t AGSCCDynamicObject::ReadInt8(const char *address, long offset)
+uint8_t AGSCCDynamicObject::ReadInt8(const char *address, intptr_t offset)
 {
     return *(uint8_t*)(address + offset);
 }
 
-int16_t AGSCCDynamicObject::ReadInt16(const char *address, long offset)
+int16_t AGSCCDynamicObject::ReadInt16(const char *address, intptr_t offset)
 {
     return *(int16_t*)(address + offset);
 }
 
-int32_t AGSCCDynamicObject::ReadInt32(const char *address, long offset)
+int32_t AGSCCDynamicObject::ReadInt32(const char *address, intptr_t offset)
 {
     return *(int32_t*)(address + offset);
 }
 
-float AGSCCDynamicObject::ReadFloat(const char *address, long offset)
+float AGSCCDynamicObject::ReadFloat(const char *address, intptr_t offset)
 {
     return *(float*)(address + offset);
 }
 
-void AGSCCDynamicObject::Write(const char *address, int offset, void *src, int size)
+void AGSCCDynamicObject::Write(const char *address, intptr_t offset, void *src, int size)
 {
     memcpy((void*)(address + offset), src, size);
 }
 
-void AGSCCDynamicObject::WriteInt8(const char *address, long offset, uint8_t val)
+void AGSCCDynamicObject::WriteInt8(const char *address, intptr_t offset, uint8_t val)
 {
     *(uint8_t*)(address + offset) = val;
 }
 
-void AGSCCDynamicObject::WriteInt16(const char *address, long offset, int16_t val)
+void AGSCCDynamicObject::WriteInt16(const char *address, intptr_t offset, int16_t val)
 {
     *(int16_t*)(address + offset) = val;
 }
 
-void AGSCCDynamicObject::WriteInt32(const char *address, long offset, int32_t val)
+void AGSCCDynamicObject::WriteInt32(const char *address, intptr_t offset, int32_t val)
 {
     *(int32_t*)(address + offset) = val;
 }
 
-void AGSCCDynamicObject::WriteFloat(const char *address, long offset, float val)
+void AGSCCDynamicObject::WriteFloat(const char *address, intptr_t offset, float val)
 {
     *(float*)(address + offset) = val;
 }
