+++ Engine/ac/dynobj/cc_dynamicobject.h
@@ -40,16 +40,16 @@ struct ICCDynamicObject {
     virtual int Serialize(const char *address, char *buffer, int bufsize) = 0;
 
     // Legacy support for reading and writing object values by their relative offset
-    virtual void    Read(const char *address, int offset, void *dest, int size) = 0;
-    virtual uint8_t ReadInt8(const char *address, long offset)                  = 0;
-    virtual int16_t ReadInt16(const char *address, long offset)                 = 0;
-    virtual int32_t ReadInt32(const char *address, long offset)                 = 0;
-    virtual float   ReadFloat(const char *address, long offset)                 = 0;
-    virtual void    Write(const char *address, int offset, void *src, int size) = 0;
-    virtual void    WriteInt8(const char *address, long offset, uint8_t val)    = 0;
-    virtual void    WriteInt16(const char *address, long offset, int16_t val)   = 0;
-    virtual void    WriteInt32(const char *address, long offset, int32_t val)   = 0;
-    virtual void    WriteFloat(const char *address, long offset, float val)     = 0;
+    virtual void    Read(const char *address, intptr_t offset, void *dest, int size) = 0;
+    virtual uint8_t ReadInt8(const char *address, intptr_t offset)                  = 0;
+    virtual int16_t ReadInt16(const char *address, intptr_t offset)                 = 0;
+    virtual int32_t ReadInt32(const char *address, intptr_t offset)                 = 0;
+    virtual float   ReadFloat(const char *address, intptr_t offset)                 = 0;
+    virtual void    Write(const char *address, intptr_t offset, void *src, int size) = 0;
+    virtual void    WriteInt8(const char *address, intptr_t offset, uint8_t val)    = 0;
+    virtual void    WriteInt16(const char *address, intptr_t offset, int16_t val)   = 0;
+    virtual void    WriteInt32(const char *address, intptr_t offset, int32_t val)   = 0;
+    virtual void    WriteFloat(const char *address, intptr_t offset, float val)     = 0;
 };
 
 struct ICCObjectReader {
@@ -63,9 +63,9 @@ struct ICCStringClass {
 extern void  ccSetStringClassImpl(ICCStringClass *theClass);
 // register a memory handle for the object and allow script
 // pointers to point to it
-extern long  ccRegisterManagedObject(const void *object, ICCDynamicObject *);
+extern int32_t ccRegisterManagedObject(const void *object, ICCDynamicObject *);
 // register a de-serialized object
-extern long  ccRegisterUnserializedObject(int index, const void *object, ICCDynamicObject *);
+extern int32_t ccRegisterUnserializedObject(int index, const void *object, ICCDynamicObject *);
 // unregister a particular object
 extern int   ccUnRegisterManagedObject(const void *object);
 // remove all registered objects
@@ -75,14 +75,14 @@ extern void  ccSerializeAllObjects(Common::DataStream *out);
 // un-serialise all objects (will remove all currently registered ones)
 extern int   ccUnserializeAllObjects(Common::DataStream *in, ICCObjectReader *callback);
 // dispose the object if RefCount==0
-extern void  ccAttemptDisposeObject(long handle);
+extern void  ccAttemptDisposeObject(int32_t handle);
 // translate between object handles and memory addresses
-extern long  ccGetObjectHandleFromAddress(const char *address);
-extern const char *ccGetObjectAddressFromHandle(long handle);
-extern void  ccGetObjectAddressAndManagerFromHandle(long handle, void *&object, ICCDynamicObject *&manager);
+extern int32_t ccGetObjectHandleFromAddress(const char *address);
+extern const char *ccGetObjectAddressFromHandle(int32_t handle);
+extern void  ccGetObjectAddressAndManagerFromHandle(int32_t handle, void *&object, ICCDynamicObject *&manager);
 
-extern int ccAddObjectReference(long handle);
-extern int ccReleaseObjectReference(long handle);
+extern int ccAddObjectReference(int32_t handle);
+extern int ccReleaseObjectReference(int32_t handle);
 
 extern ICCStringClass *stringClassImpl;
